vendor-merge: Adaptations to MPICH 4.
This patch implements adaptation of the PSP device to the MPICH 4 code
base. These include among others:

* coll
  - Promote count types (MPI_Aint) for large count support
* comm
  - Add support for  `MPI_COMM_TYPE_HW_GUIDED` and
    `MPI_COMM_TYPE_HW_UNGUIDED`
* configure
  - Fix handling of fortran calls with mismatched arguments
  - Fix handling of `--with-hwloc`/`--without-hwloc`
* dist
  - Fix non-CUDA-aware RPM builds
  - Remove obsolete dependency to pscom for RPM-CH4 builds
* group
  - Adapt lpid to `uint64_t` given by the `MPIR_Group` struct
* logging
  - Adapt to simplified MPIR-level logging
* node ID
  - Disable use of MPID_Get_max_node_id (for the non-MSA case)

+++ mpich2/src/mpid/psp/include/mpidpost.h
@@ -149,16 +149,16 @@ static inline int MPID_Progress_test(MPID_Progress_state * state) /* state is un
 struct MPIR_Comm;
 int MPIDI_GPID_GetAllInComm(MPIR_Comm *comm_ptr, int local_size,
 			   MPIDI_Gpid local_gpids[], int *singlePG);
-int MPIDI_GPID_ToLpidArray(int size, MPIDI_Gpid gpid[], int lpid[]);
+int MPIDI_GPID_ToLpidArray(int size, MPIDI_Gpid gpid[], uint64_t lpid[]);
 int MPID_Create_intercomm_from_lpids(MPIR_Comm *newcomm_ptr,
-			   int size, const int lpids[]);
+			   int size, const uint64_t lpids[]);
 int MPIDI_PG_ForwardPGInfo( MPIR_Comm *peer_ptr, MPIR_Comm *comm_ptr,
 			   int nPGids, const MPIDI_Gpid gpids[],
 			   int root, int remote_leader, int cts_tag,
 			   pscom_connection_t *con, char *all_ports, pscom_socket_t *pscom_socket );
 int MPID_Intercomm_exchange_map( MPIR_Comm *local_comm_ptr, int local_leader,
                                  MPIR_Comm *peer_comm_ptr, int remote_leader,
-                                 int *remote_size, int **remote_lpids,
+                                 int *remote_size, uint64_t **remote_lpids,
                                  int *is_low_group);
 
 int MPIDI_GPID_Get(MPIR_Comm *comm_ptr, int rank, MPIDI_Gpid gpid[]);
